diff: create package to generate unified context diffs
The current format for diff output in muscle diff and muscle history
commands is non-standard and utterly untested. Here we implement a format
that's compatible with GNU diff (unified context diff), and we implement
the diff routines with decent testing.

Note that one could use snapshotsfs and muscle history to compare files
in different revisions of the file system, e.g.,

	diff -U 3 /n/snapshots/$rev1/dir/file /n/snapshots/$rev2/dir/file

but for filesystem-wide diffs we need to use the Merkle tree to trim
the recursive diff processing.