tree: add instance to NewRevision and adjust calling code
New revisions are created by tree.Tree, while led me to adding an instance
property to the tree.Tree type. A tree belongs to an instance. A tree's
instance will usually be set from the revision that the tree corresponds
to when it is created. In some cases, since old revisions have empty
instance property, and because an initial revision may not exist when
first getting started with muscle, we've added a special "constructor"
NewTreeForInstance that sets the instance explicitly (will use the value
in lib/muscle/config).

The tree.Store.Fork method sets the right instance in the target revision.