cmd/musclefs: after a snapshot, make the local revision a child of the pushed remote revision
During routine merging on my instances I noticed that I was doing a little more
conflict resolution than expected, which, as I saw via the generated diagrams, was
due to a non-optimal merge base being used. That in turn, is due to a bug where
the local revision was not correctly re-parented after pushing the remote revision
(there are more details in the comment added to the code).