tree: always show diff heading when one of the two nodes is nil or the type has changed
At some point I checked out a large repository, and when taking a snapshot of the fs,
I obtained an unexpected inordinate amount of blobs to upload. (I forgot about the large
repository.)

Turns out muscle diff (or muscle history -d) would not show addition of a new directory.
In terms of content, both nodes would look empty (nil file node is empty, directory node
has empty content).

With this change we always show the header like the following

	--- /dev/null
	+++ b/root/src/plan9front

when one of the two nodes is different, or the file type has changed.