tree: remove the notion of local revisions
This is a breaking change, in that this commit won't work with the
current version of the root key file at $home/lib/muscle/root.

The root file $home/lib/muscle/root will now contain the pointer to a
root node, not to a revision.

A fictitious local revision is made up from the current root node,
having as parent the current remote revision. This makes history and
merge commands work.

Keeping local revisions was almost pointless anyway, because when taking
a snapshot all the local revisions were lost. It's true that they provided
fine-grained recent local history, but I've seldom used it.

This is a first step towards further simplification of the staging area.