all: migrate to new internal/block package for node blocks
The aim of this change is to encapsulate reading/writing from storage
into an internal/block package that could at some point be changed so
that the index doesn't use files on a file system but a block device
directly. A more immediate goal is to have less churn in the index
(many more files are created there than are necessary.)

It would probably be easier to rewrite the software at this point, but
evolving it is more instructive (plus, the software already works for me,
this is just experimentation).

More changes need to be done in the future so that node and revision
metadata are written using internal/block rather than using the storage
package directly.

Change-Id: Iff99d164fce70228bf9e4b682541edb7381b501f