tree: use loaded flag instead of implicit condition
A Node without a name is considered a Node not loaded from storage,
a Node of which only the key is known. Although I see no use for a Node
without a name (apart from the root), it seems cleaner to use the flags
property of Node for this.

I've looked for comparisons of Name with "" to validate this,
and assignments of "" to Name.

Change-Id: Id36b390c18d1bf726ce2a1d645a2aa6e1ea164bb