tree: use DefaultBlockSize only for new nodes
The usages of tree.DefaultBlockSize have been replaced by tree.Node.bsize.
Such property is properly populated by node.Tree.Add (for new nodes) or
by the codecs when deserializing (existing nodes). Both v13 and v14 treat
0 as the default block size before this change, which is 1024*1024 bytes.

After this change is merged, it will be possible to change the value of
the tree.DefaultBlockSize constant, affecting the block size for nodes
created after such change. Further, it is possible to make it a variable
bound to configuration.

Change-Id: I28115209e9f81c6b86c75b9719ec887f447f24d2