all: add capacity argument to internal/block.Factory.New
Why this is necessary: If the block factory decides the capacity of all
blocks, all nodes in the system, old and new, must have the same capacity;
that setting can't be made configurable. Therefore it's up to the client
to specify the capacity for blocks.

This is still not great  because it seems strange that
internal/block.Factory.New would need to be told what capacity an
*existing* block should have. That's a problem for another day.

Change-Id: Ie95b28d80058e45c1108abf693021efd091521d9