tree: use internal/block to store revisions
This avoids using the storage package for loading/storing revisions.
After this change is merged, only tree.Node requires direct access to
storage.  The idea is to also store tree.Node's using internal/block,
so that we can then refactor internal/block to use a block device for
the index, and that will be transparent.

The way we store/load revisions via internal/block is
convoluted, to avoid changing the type of tree.Revision.key, which would
make the CL quite extensive.

Change-Id: I51853adb9ec3ced0f76e95682c7a73f3a156c5fd