tree: do not try to automatically handle data corruption scenarios
(1) We used to add empty nodes named 'vanished' in case a block isn't
found in storage.

(2) We used to add empty nodes named 'nocodec' in case a block
required a codec that muscle no longer supports.

(3) We used to rename child nodes so that their names are unique (in
particular, this would work with (1) and (2) in case more than one
node vanished or was unreadable).

Today I don't think any of those strategies help. If the data is
corrupted, it's got to be fixed. Specifically:

(1) If a blob is not found, try to find a machine that has the
missing blob. Hopefully it still is in some cache. If not, use
"muscle control unlink path/to/node" (data loss).

(2) Add back support for the codec!

(3) Rename/remove one by one (walk will return only the first match).