cmd/musclefs: add “diff” control command
Allows to say "muscle control diff" (or "muco diff", if you use an
alias) rather than "muscle diff". The latter may fail if not run
right after the fs has been flushed (which happens periodically, but
can also be forced with "muco flush"). At the moment I don't see a
use for "muscle diff" but will leave it in the repository for now. It
may be useful to diff two arbitrary revisions (but it would need to
be slightly extended).