cmd/musclefs: implement DMEXCL
Like CWFS, we'll allow setting the DMEXCL flag on a file that's
already open more than once.

I chose to implement DMEXCL as a feature of musclefs, not as a
feature of the muscle data model. As a consequence, I'm tracking
extra mode bits (only DMEXCL at the moment) per file system node.

This implementation passes the basic tests that I wrote to explore
the reference implementation (CWFS) but there are almost certainly
edge cases (esp. concurrency wise) that I haven't considered.