storage,cmd/musclefs: nudge the paired store to start uploads (fsdiff)
One of the properties tested by fsdiff is that it's safe to remove
the local cache entirely once the data has been propagated to the
remote store—this happens after a "push" command, and completes
asynchronously. In order not to have to wait for seconds in fsdiff,
I added a mechanism to wake up the goroutine that uploads blocks
after a "push" command. This actually make the polling entirely
unnecessary, so it's been removed.