tree: make it explicit that there can't be half-renames (validated via fsdiff)
Rename consists of two operations: (1) unlink a node from its
parent, and (2) link the node to its new parent. Both must happen,
or neither. Otherwise data is lost (or duplicated). The code showed
some error handling suggesting that (1) could happen without (2) to
follow. I inlined the called functions' bodies and simplified them
for the case in question. Note that data loss can still happen if
the code panics or some invariant is violated... trusting fsdiff to
validate that this code works in practice.