cmd/musclefs: allow read/write from unlinked node (fsdiff)
Testing musclefs against ext4 with fsdiff, showed the following bug:
Removing a file via unlink(2) or Tremove immediately removes the
underlying blocks. This breaks reads/writes via fds/fids referring
to the unlinked node.

I wrote two minimal reproducers for the bug:
* create, write, close, open, unlink, seek, read; and
* create, write, close, open, unlink, seek, write, seek, read.

The second version has a write after the unlink. I think this brings
back the discarded block as a zero block. So the read returns 8
bytes, but the first 4 are zeros.

The first version doesn't have a write after the unlink. In that
case, the block is discarded and the read returns 0 bytes.

The node should really only be removed on Tclunk or Tremove, if no
other refs remain. It should only be marked as unlinked.

According to remove(9p) though, musclefs's implementation is a
little more aligned with Plan 9's. Not quite though, because the
write should return a phase error rather than succeed.

> If a file has been opened as multiple fids, possibly on dif-
> ferent connections, and one fid is used to remove the file,
> whether the other fids continue to provide access to the
> file is implementation-defined.  The Plan 9 file servers
> remove the file immediately: attempts to use the other fids
> will yield a ``phase error.''  U9fs follows the semantics of
> the underlying Unix file system, so other fids typically
> remain usable.

I chose to align musclefs with ext4 (because I most often mount
musclefs in Linux), deferring discarding the file until the last
reference to it is gone.

The removal of checks in the Tread/Twrite handler makes grafting
nodes unsafe (because now we don't prevent i/o to unlinked nodes).
I'll address that when I add pull/push commands to fsdiff.