cmd/musclefs: fix rename via wstat (fsdiff)
Musclefs has two rename mechanism. One is via the control file:

	echo rename x/y z/w >ctl

which works like rename(2). The other is natively supported by 9P in
Twstat, and is used when a file or directory changes name within the
same directory it already belongs to. In the Twstat implementation,
there were two problems:

* it allowed overwriting non-empty directories,
* it left orphaned blocks in the staging area in case of overwrite.

This is now fixed. But note, stat(5) says:

	The wstat request can change some of the file status infor-
	mation.  The name can be changed by anyone with write per-
	mission in the parent directory; it is an error to change
	the name to that of an existing file.

Therefore, allowing overwrites in Twstat is a deviation from 9P.
As noted in the code, this is necessary to make use of programs
that expect such overwrites to work (e.g., git) when the fs is
mounted via 9pfuse or the Linux 9p driver.