internal/tree: only rename2 when source and target have no open refs (fsdiff)
Suppose a source node "/november" has a fid pointing to it and that
is moved via rename2 to "/mike" (note, Wstat would suffice in this
case). Clients may want to clone the fid and then Tremove, thinking
they'll remove "/november", but they'll remove "/mike" instead.

Similarly, suppose a target node "/mike" has a fid pointing to it,
and we rename2 "/november" to "/mike". The client may deny creating
"/mike" because of this fid, though it points to "/november" now.

Therefore, I chose to disallow rename2 when source or target has a
reference (including via any of their descendents).

I thought of updating the fid to node mapping upon doing a rename2,
but the usefulness of allowing that doesn't grant the additional
complexity, and, if the project history is any indication, the
generation of additional edge cases.