internal/tree: do not graft over nodes with references
The reasoning for this is the same as that for disallowing renames
where source or target have references. It was allowed in the past
because I disallowed all operations to unlinked nodes, invalidating
them. To get POSIX compliance (admittedly a weird thing for a 9P file
system) I now allow such operations, but then it's unsafe to graft
over nodes that are referenced by some fid.