internal/tree: must make sure node is loaded before adding a child
Not doing so opens up the possibility of adding a child with a name already taken.
Noticed this writing a bespoke 9P application, rather than interacting with the fs
via a mountpoint. My code does a Tcreate and falls back to Topen if the file already
exists. Doing the Topen first, or a Tstat, woud load the node—I expect this is what
the fs driver does.