all: support arbitrary tags for each revision
Musclefs supports tagging sub-sequences of revisions; these could be
useful for tracking projects histories as sub-sequences of the whole
fs history. The commands for diff and history all support a new `-b`
option (tag to use as base to diff from, show history of revisions
with given tag only). The push command, used to create new revisions,
supports an optional list of additional tags (in addition to the
default, "base") to add to the revision.